64-bit Porting. JNI pointers passed in int.
BUG=354405, 346626

Review URL: https://codereview.chromium.org/263833002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267866 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 02bbad5b516e9ab3c624d40f4c95e1dbcbb5cd91

+++ chrome/android/java/src/org/chromium/chrome/browser/TtsPlatformImpl.java
@@ -132,7 +132,7 @@ class TtsPlatformImpl {
      * @param context The app context.
      */
     @CalledByNative
-    private static TtsPlatformImpl create(int nativeTtsPlatformImplAndroid,
+    private static TtsPlatformImpl create(long nativeTtsPlatformImplAndroid,
                                           Context context) {
         return new TtsPlatformImpl(nativeTtsPlatformImplAndroid, context);
     }
